Update patch set 1
Patch Set 1:

Chenfei, FYI this is your first submit. Check the logs and then submit a patch. Looks like the JUnit test for guard failed.

Jorge, please feel free to comment on how the code was restructured. I want to make sure the packages directory is correct and that we get the archetype correct also.

Patch-set: 1