Update patch set 1
Patch Set 1:

Because of the renaming of onap in drools-apps, we need counterpart changes in the docker side to pull in the correct apps zip.   

Can you submit counterpart changes in docker, the root pom.xml, search for maven-dependency-plugin, and change coordinates for the apps zip package.  I think that's the only change that is needed there in order to have their installation automated.

Let's try to submit both patches at the same time to avoid integration problems when using the staging version, as there are integration testing activities going on with all components.

Other than that, looks good, I'll approve once I see the docker review with that change.

Patch-set: 1