Update patch set 2
Patch Set 2:

Pam, this patch set 1 may not be the latest. I have ignored the unit test for template.demo due to an internal database access by guard which will cause the test to fail. The following build failure is still due to the test failure in template.demo. So I think this is not the latest version. The latest version is in my newer commit. This commit is the original one.

Patch-set: 2
Reviewer: Gerrit User 1441 <1441@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0