Update patch set 1
Patch Set 1:

Jorge, if I get it right, we used a DB in research network to store the operations to support guard functions. For ONAP, we need to change to use a in-memory DB. I think we have a separate ticket for the changes to use in-memory DB. Michael will commit once its done.

Patch-set: 1