Update patch set 4
Patch Set 4: Code-Review-1

(17 comments)

The more sever issue and question is the addition of "transient" keyword to some fields.   I think this may be incompatible with persistence of facts in kie sessions.   In addition, if the fields are serialized in json format, I wonder if it does serialize those fields.

The other comment is the inconsistency of output, we have some logs using slf4j and others System.out/System.err - they may end up in different files.  I think all should move together, even though is more work for you.

Can you verify that junits are not missing any of the fields marked as transient?

Patch-set: 4
Label: Code-Review=-1