Update patch set 2
Patch Set 2:

(2 comments)

General comment, I notice in the logger statements calling explicitly obj.toString().   It would be better to just pass the object directly and it will print null or the toString() method automatically.  Otherwise I think if the "obj" is null it could throw a nullpointerexception because of the logging statement

Patch-set: 2
Label: Code-Review=0