Fix JUnit Test for ControlLoops
Fixed the test case by using getters and setters for Policy objects
in the drl. Added gson dependency in the pom to fix gson runtime errors.

Fixed dependencies in template.demo pom to add vfc.

The drl was having compilation errors in the guard permitted
rule due to the operation request object not being casted
into a VFCRequest when being passed into the VFCManager.

Issue-ID: POLICY-194
Change-Id: Id71c74ad9727ad942a683f63a0dfad3f22b6376b
Signed-off-by: Daniel Cruz <dc443y@att.com>