CL dependencies fixes for lab environment runtime
This also includes workarounds to the recent oparent dependency
introduction that breaks runtime (with the version-check-maven-plugin).

manifested by loading control loops and failing to load some classes
due to different versions.

The issue was that underlying drools libraries use 3.2.5 and oparent
has included a had dependency with transitive dependencies for some maven
libraries in 3.2.3 and lower version xml parsers.   Bottomoline, the
classpath at runtime was formed by the union of both, with some
libraries being resolved to the oparent one, and others to the drools
one.   These errors are very obscured to debug.

Additional clean up of dependencies versions and order of build
was introduced to avoid issues loading dependencies at runtime in a
lab environment (non-junit)..

Issue-ID: POLICY-162
Change-Id: I019c82e6bed4eab4884cdbf8f6f32472c3a7352f
Signed-off-by: Jorge Hernandez <jh1730@att.com>