Update patch set 2
Patch Set 2:

(7 comments)

Jorge, the reason java code is changed more than junit code is that the drl calls the "production" code, the junit does not. So there is no good unobtrusive way to specify that it is running a test as far as I know. There can be a meeting to clarify all the concerns.

Patch-set: 2