Update patch set 3
Patch Set 3:

Could someone please look at the test case that failed? It happens intermittently. Sometimes it works, sometimes it fails. I'm sure it is a timing issue for one of the Thread.sleep() calls. If someone wants to own that then we should file a JIRA for it.

Patch-set: 3
Label: Code-Review=0