Fix A&AI Named Query for vFW
The vFW named query has now been verified with the
A&AI simulator. Null checks were added to make the
namedQuery method work properly. In addition, a
recursive A&AI response parser was added to traverse
the response to find the correct generic-vnf object
with the target vnf-id.

For now the simulator URL is hard coded and the UUID
of the named query is hard coded. These should be
configurable in future releases.

Issue-ID: POLICY-104
Change-Id: I05a1a992ff68ca2c17fb6a578983e6b21626bf44
Signed-off-by: Daniel Cruz <dc443y@att.com>