Update patch set 2
Patch Set 2:

> Thanks Daniel, good work.  I think we talked error case handling
 > will come later, case, aai query fails or don't get enough info,
 > correct?

 > Thanks Daniel, good work.  I think we talked error case handling
 > will come later, case, aai query fails or don't get enough info,
 > correct?

Right, for now what happens is if it does not find the proper vnf-id it well be nulled out and not sent in the request but the request is still sent. This will result in APPC giving us a failure in which Policy will send an operation failure and then a final failure. In the future we should not send a request if any information is missing and retract everything from memory in my opinion. This will be for all actors, not just APPC.

Patch-set: 2
Reviewer: Gerrit User 1677 <1677@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0