Update patch set 3
Patch Set 3:

I think it is the intermittent failures since patch2 and 3 are logically the same. But on the other hand, it also looks like the the build is failing on VfcSimulatorTest. It seems like it fails to connect to the DB: Connect to localhost:6668 [localhost/127.0.0.1] failed: Connection refused (Connection refused)

I don't think my changes touch VfcSimulatorTest at all.

Patch-set: 3
Label: Code-Review=0