Fix Use Case Testing
These changes are necessary to get the use cases working
on an actual PDP. So far vCPE, vFW, and vDNS have been tested
successfully on a pdp. These are priliminary tests to just
test a FINAL SUCCESS without A&AI GET queries being used.
A&AI named queries for vFW and vDNS have been verified to
work.

Changes in the template were made for SO as there were
changes in the JUnit template that did not get reflected
in the archetype template.

Changes in the code were made in various places to use
the environment properties for REST urls, usernames, and
passwords. Due to VoLTE not being tested yet, the url
for VFC is still hard coded and will be changed later.

It has been confirmed with APPC that LCM will have two
topics: APPC-LCM-READ and APPC-LCM-WRITE. Policy needs
to sink to APPC-LCM-READ and pull from source
APPC-LCM-WRITE. This has been reflected in the controller
properties file.

Issue-ID: POLICY-259
Change-Id: Ib9a8df07ae5ad9d3052c88907c1e522952af474d
Signed-off-by: Daniel Cruz <dc443y@att.com>