Update patch set 1
Patch Set 1:

Hi Ritu, trying to get as much code as possible merged in so to testing in lab, we are running out of days, and would like to have as much as possible as close to the final product early tomorrow morning.   Let's see what Daniel answers.  The code seems correct, just cosmetics.

Let's work out any changes in another review, if agreement between you and Daniel is reached with regards to the changes you are questioning.

Patch-set: 1