Update patch set 1
Patch Set 1:

I'm fine with deleting this since the amsterdam code base doesn't use it but one concern is that the legacy template does use this package. Do we need to keep it for since we are keeping the legacy template or will that be removed soon as well?

Patch-set: 1