Update patch set 1
Patch Set 1:

> > > > > looks good to you Ali?
 > > > >
 > > > > Yes, but it seems redundant as there is already code in the
 > > > > VFCManager that sends a restful request in a while loop
 > (while
 > > > > (attemptsLeft-- > 0 where attemptsLeft is hardcoded to 20).
 > > > Please
 > > > > see line 92 in VFCManager.java
 > > >
 > > > Hi Ali, That is the case when VFC sends HTTP response
 > > accepted(202)
 > > > and the code is waiting for the request to complete and send
 > > > sucess(200). This is one request only.
 > >
 > > Hi Ritu,
 > > So what if you get a failure response (i.e 400 bad request)? Then
 > > it will retry the request up to 20 times until you get a success,
 > > correct?
 > 
 > Yes that is correct. The attempts here are waiting for one request
 > to be complete and not sending new requests. From VFC perspective
 > only one request was sent.

Right, and that is the same functionality as the code change you made in ControlLoopOperationManager which is why I think it is redundant. I suggest you keep 1 of these - either get rid of the change you made to ControlLoopOperationManager and keep the while loop in VFCManager and maybe instead of hardcoding to 20 attempts you can grab the 'retry' value from the yaml OR keep the change you made in ControlLoopOperationManager and get rid of the while loop in VFCManager (send the request without while loop). Jorge, what do you think?

Patch-set: 1