Update patch set 1
Patch Set 1:

> > > > > looks good to you Ali?
 > > > >
 > > > > Yes, but it seems redundant as there is already code in the
 > > > > VFCManager that sends a restful request in a while loop
 > (while
 > > > > (attemptsLeft-- > 0 where attemptsLeft is hardcoded to 20).
 > > > Please
 > > > > see line 92 in VFCManager.java
 > > >
 > > > Hi Ali, That is the case when VFC sends HTTP response
 > > accepted(202)
 > > > and the code is waiting for the request to complete and send
 > > > sucess(200). This is one request only.
 > >
 > > Hi Ritu,
 > > So what if you get a failure response (i.e 400 bad request)? Then
 > > it will retry the request up to 20 times until you get a success,
 > > correct?
 > 
 > Yes that is correct. The attempts here are waiting for one request
 > to be complete and not sending new requests. From VFC perspective
 > only one request was sent.

Hi Ali, In my understanding these two are different things. In VFCManger we are waiting for VFC to finish completing the request. VFC may not be able to complete the request right away. This is for one request. The REST call in the loop here is a GET call based on the jobid received from a POST call made before outside the loop. If the request fails VFC is not going to retry it based on the GET call but keep returning the same result. The YAML file retries is for actually trying again by sending another POST call.

Patch-set: 1
Reviewer: Gerrit User 285 <285@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0