Update patch set 2
Patch Set 2: Code-Review-1

Michael is in PTO today and tomorrow. My only concern is if there is a guard deny and another operational policy chained to failure_guard, can it be proceeded? I roughly remember that was the reason why we should startOperation before querying guard. Otherwise, if we get a guard deny, finishOperation will not really work since there is no currentOperation assigned before. Thus, it will just throw an exception saying no operation to finish and not proceed. Daniel, could you check that?

Patch-set: 2