Add hot fix for guard db on disable
Added a hot and inefficient fix so that the guard database is not
attempted to be accessed while guard is set to disabled. Added booleans
in the drl files to manually set guard to be disabled. The guard db
access is done in the operations manager which has no context of guard
at the moment. Because of this separation, the booleans must be passed
into the operation manager to notify that guard is disabled and not to
connect to the db. This is a symptom of a need for guard refactoring.

Issue-ID: POLICY-260
Change-Id: Ida7a2a7c19e568b076731e0f04bdc8a9fd0fcb26
Signed-off-by: Temoc Rodriguez <cr056n@att.com>