Add fix for guard db on disable
Added a fix so that the guard database is not attempted to be accessed
while guard is set to disabled. The guard db access is done in the
operations manager which has no context of guardat the moment. Because
of this separation, the boolean must be grabbed from the PolicyEngine
properties. We should consider some guard refactoring.

Issue-ID: POLICY-296
Change-Id: Ida7a2a7c19e568b076731e0f04bdc8a9fd0fcb26
Signed-off-by: Temoc Rodriguez <cr056n@att.com>