add structure only for PAP rule management
PAP currently generates a set up rule that inserts a Params object
when the session is loaded.   Since the official template also
has an equivalent rule that is used by the drools applications
team to perform initializations tasks, at runtime we end up with
2 Params objects with the same control loop data.   This has
sideeffects where the same rule becomes true for 2 Params objects
therefore performing duplicated actions which introduces somewhat
anomalous behavior.

The PapParams is introduced to satisfy current PAP needs for
management of rules.  The structure is included here only for
those purposes but it carries no significance at runtime during
drools executions since the rules go by Params instead of PapParams.

In the near future, the plan is to have the PAP not generating
SET UP rules.

There is a companion review on the PAP side to refer to PapParams
instead of Params to fulfill the resolution of POLICY-302.

Change-Id: I2f9ed3cee3d5d5209eecd43823cfc44a4b937600
Issue-ID: POLICY-302
Signed-off-by: Jorge Hernandez <jh1730@att.com>