Fix operation history DB properties for Junits
Renamed operation_history.properties to operations-history.properties.
Changed the properties in persistence.xml and operation_history.properties
so that they use the correct values. Previously junit would never write
to db because the "guard.disabled" property was not set. Set this property.
Additionally the operationshistory10 table was trying to be created even
though it exists. These values work in junit code but there is still the main
issue of not writing to the DB in production. Could not test on
rackspace yet. Changed the keys in the operations-history.property.

Issue-ID: POLICY-345
Change-Id: Id755b3152ae6ac4991ea9c314154592ea0105070
Signed-off-by: Temoc Rodriguez <cr056n@att.com>