Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

I am not clear, if the username/password values are hardcoded in the *.properties from the review.  If it is the case, it is not acceptable for production.  It does not reconcile with the way that those values have been used from the control loop environment properties that is used in other guard logic.

Patch-set: 1
Label: Code-Review=-1