change drl line to avoid velocity bug
generating a control loop project from archetype
(but not from junits) runs into velocity parse failures
as specified in ticket.   The existing .drl code
for the line that fails the parsing has been slightly
modified to bypass this problem.   In addition a
logging statement has been added to the catch.

Change-Id: I3eec82c9b710cc9d2a13dc68032a8a6074016b40
Issue-ID: POLICY-376
Signed-off-by: Jorge Hernandez <jh1730@att.com>