Update patch set 1
Patch Set 1:

We definitely retract them. We could publish a Reject message so folks know what we did with it. Ideally (future), there would be a policy that dictated what to do but I think for now we just publish a reject. The requirement really should be on DCAE not to do this behavior.

Patch-set: 1
Label: Verified=0