Write to DB on guard deny
When an operations was denied by guard, it was not writing to the DB at
all. Added an else-if to the ControlLoopOperationManager to handle a
guard response. This is now called from the guard response rule only in
the deny case since we want to finishe the operation at that point. Had
to add start operation just before finishing it. This allows the write
to occur.

Issue-ID: POLICY-345
Change-Id: I665b9cfaa684493e39e731102d4d43ecb39fad7c
Signed-off-by: Temoc Rodriguez <cr056n@att.com>