Update patch set 1
Patch Set 1:

At this point in the release, it is critical not to introduce any undesired side effects, goal is to release next week, and that is tested against all 4 use cases with guard and without guard enabled.  One question that comes to mind if there are side effects with guard disabled on the .drl of new changes?

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0