Update patch set 1
Patch Set 1:

That is a good question Jorge. The only possible way this new code runs is if the connection to pdp-x is successful and it returns deny. So it will not run on guard disabled. But when it does run, it starts the operation, sets response, writes to db, adds the current operation to the operationHistory linkedlist, and finishes the operation. Those are the areas where we should be concerned if there are side effects, if any. It does the same stuff as the other 'response handlers' so it should be ok. Additionally, if any clean up / tear-down rules in the drl need to be run, I am not 100% about them occurring after the operation is denied since this behavior is brand new.

Patch-set: 1