Update patch set 2
Patch Set 2: Code-Review+1

Looks fine. In the future we may consider refactoring this. I think startOperation is taking care of too many tasks. It would be better to separate the logic where the requests are retried and constructed (this should stay in the startOperation method) and the setting of the operation (should be moved to it's own function). If we have setting the operation in its own method we could call it in the constructor of the operation manager so that by the time we need to check the current operation in the case of a guard deny, it will not be null and start operation does not have to be called twice. This is ideal as the intent of startOperation is for actors specifically, it becomes confusing if we start adding in logic to handle non actors.

Patch-set: 2
Reviewer: Gerrit User 1677 <1677@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1