Add a DMaaP simulator
Added a DMaaP simulator for testing applications.  One current
limitation is that the simulator does not support multiple "subscribers"
on the same topic; if someone gets a message, that message cannot be
subsequently retrieved by anyone else.  The simulator has also not been
tested for concurrent getting and posting.

Also added a way to set the response code DMaaP would return for a get
to test that policy can gracefully handle errors.

Issue-ID: POLICY-489
Change-Id: I0fa559a544f440ec5ad2360209899185e253ae34
Signed-off-by: Charles Cole <cc847m@att.com>