Update patch set 4
Patch Set 4: Code-Review-1

Good work Charles!   Some feedback, since ueb, and dmaap are part of the drools-pdp infrastructure.   It is best to move just this simulator under the drools-pdp.

I expect policy-endpoints in drools-pdp junits beefed up considerably now that there would be simulator.   So, a subsequent step after this work, would be to beef up the junits in the drools-pdp side.

But drools-pdp should not depend on drools-applications, therefore this work is best to be moved in the drools-pdp repo.

Ideally this can be a simulators feature in the drools pdp.   Hopefully the feature can be used both ways: a) simulator, and b) junits.  One consideration is then circular dependencies, I wonder if that could be avoided setting the maven-surefire-plugin classpath entries.

Patch-set: 4
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1