Update patch set 2
Patch Set 2:

Hi, the reason of this change is the following.

AAI comes with a certificate that has a CN being aai.simpledemo.onap.org. But in OOM, the hostname of AAI is aai-service.onap-aai hence the hostname validation failure.

To solve this issue, we could levrage Subject Alternative Name in the certificate, but that wouldn't work well neither. Because in OOM, the goal is to be able to deploy ONAP multiple time, and a deployment is segregated using its namespace. Basically, for aai it would be aai-service.<namespace>-aai
So as you see, hostname is very dynamic.

Finally, the overall certificate mgnt issue in ONAP beings other requirement, like the ability to substitute ONAP's certs by proprietaries' ones, e.g. when you go on production with ONAP, you want to reference your own certs, not the ONAP's one.

I understand the security risk here, but also, this change in the RESTManager is not something exposed to the external world, it lives in the internal of ONAP. e.g. only an app (SO/VF-C/...) triggering something in policy can trigger the RESTManager functions (I think).

Patch-set: 2
Reviewer: Gerrit User 137 <137@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0