Update patch set 2
Patch Set 2: Code-Review+1

I think we will have to do some changes to disable the default behavior of disabling hostname validation in the cert, but I'm ok with it for the time being, if everybody else is.  I can open a jira ticket to track it if Pam and Liam are okay with it.

Patch-set: 2
Label: Code-Review=+1