Update patch set 1
Patch Set 1:

The asynchronous REST call was there in the original code. I turned the anonymous inner thread class into the AsyncSORestCallThread private class in the SOManager becasue SONAR moaned about it for tehcnical debt.

I added a return value from the existing asyncSORestCall() method so that in the unit test we can monitor execution of the thread. Before, this method had a void return.

The core in the run() method of AsyncSORestCallThread is functionally the same as what was there before in the anonymous inner class.

Should that code check for 202? If so then we need to modify the code to do that.

BTW, I could not see where this code was called from or used either in the Java classes or the Drools rules.

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0