Fix package directory naming to lower case
In macOS (and windows) directory names are case insensitive, but in
Linux they are case sensitive. Therefore, when the "actorServiceProvider"
directory name was renamed to "actorserviceprovider", the change did not
propogate into git when a Macbook was used for development. This error
was discovered when working on a Ubuntu machine and is now fixed.

Signed-off-by: liamfallon <liam.fallon@ericsson.com>
Issue-ID: POLICY-455

Change-Id: Ifebe3d37d42e79fff8da2370369967a25c371b11
Signed-off-by: liamfallon <liam.fallon@ericsson.com>