Tidy dependencies and scopes in POMs
There were some duplicate dependencies in the POMs that were causing warnings.
These have been removed. In cases where the revision of the duplicate
dependencies were different,the latest revision was used.

The scope of the "junit" dependency was "provided" in some POMs. These have been
corrected with the scope of "test" being used for junit dependencies.

Issue-ID: POLICY-238

Change-Id: Ib1cc6f040b0c4d7f2b39ae02bc6fb4590d395809
Signed-off-by: liamfallon <liam.fallon@ericsson.com>