Update patch set 4
Patch Set 4:

Hi Jorge,

I didn't realise we were permitted to play with the group ID. In Ericsson internal work the group IDs are fixed and we can't change them so in ONAP we have more freedom!

I agree, it would be much better to work with he group IDs rather than the artifact IDs because then the structure would be reflected in the Nexus repo as well. I'll modify the structure for that.

I didn't understand your comment about omitting the reactor at some levels; was this that we would jump down a certain level of hierarchies in the directory structure ignoring intermediate levels?

On side effects, with these sort of changes I agree that doing a full build passing unit tests is not enough. We need to do a full build of the entire policy framework, and deploy it on a production-like environment to verify that we have no problems. The build and deployment here is complex so it would be a miracle if we got all the changes perfectly correct first time. However, it's pretty easy to make the changes so after a few tries I think we'll get it right.

Patch-set: 4