Update patch set 1
Patch Set 1:

When this JIRA was assigned to me I changed the oparent version on a clean clone and tried the build, which worked. I'm afraid I didn't do any deep analysis as I thought I'd try it first and see what happened.

The only snapshot reference in this repo is the one below.
/controlloop/common/msb/pom.xml
    <dependency>
      <groupId>org.onap.msb.java-sdk</groupId>
      <artifactId>msb-java-sdk</artifactId>
      <version>1.1.0-SNAPSHOT</version>
    </dependency>

I see that there is a 1.1.0 version of that dependency in the ONAP staging nexus now so we cold change that one. I just changed it on my local copy and ran a build which worked. We should change that dependency.

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0