Update patch set 2
Patch Set 2:

Jorge, we do not need a work item to be created. This version of oparent will fail in the release jobs if there is a SNAPSHOT version. But of course we will be follow the guidelines as directed by ONAP versioning and release process. I sent you this previously.

Patch-set: 2