Fix Serialized Names of A&AI Subtag
Onset events received from DCAE contain an "AAI" subtag that is
capitalized. The casing of the AAI variable was changed to lower case which was making the gson serializer unable to deserialze a simulated onset event that
matches the casing DCAE uses, which resulted in Policy rejecting the
events because they appear to be missing "AAI".

A serialized name was added to preserve the lower cased variable as
sonar wants but will also allow onsets with capitalized AAI
subtags to be processed by the rules engine.

Issue-ID: POLICY-681
Change-Id: Ia37694f606b3fb0898d6c4dadea7c48c5372e64e
Signed-off-by: dc443y <dc443y@att.com>