Create archetype for Beijing Release
This new archetype is a proof of concept for removing
template expansion from the BRMSGW. The idea is to have
a ueb or dmaap topic that the BRMSGW uses to send params
json and then the pdp-d will deserialize it into a
ControlLoopParams object. The rules engine will then
insert that fact in to the working memory which will be
used for rules matching between onset events and
corresponding policies.

Issue-ID: POLICY-692
Change-Id: I4237221c2216dab16160e79a8f2fdcab6571e1eb
Signed-off-by: Daniel Cruz <dc443y@att.com>