Update patch set 1
Patch Set 1:

(1 comment)

Two more things to note: This is just a proof of concept so it doesn't include a solution for persisting the ControlLoopParams objects when the sessions go down. Also for active/active there needs to be a way for onset events that are coming in to go to the PDP that contains the corresponding ControlLoopParams object.

Patch-set: 1
Label: Code-Review=0