Create Drools Scripts for Beijing
The create-cl-beijing script will now allow the user to create a rules
jar that does not need to be expanded and also generate test files that
can be used to insert facts for all the supported beijing use cases. The
yaml for each policy is provisioned by the user in case custom yaml is
desired.

In addition to this, a new script is included that will extract the
needed data from the controller properties file to insert facts for all
the supported use cases. This script can be used to insert
ControlLoopParams facts after the create-cl-beijing script is used to
generate the rules artifact. No provisioning necessary other than
specifying where the controller properties file is located.

Issue-ID: POLICY-692
Change-Id: I42f0a08fba133ca36fb1be588a720e4f9598d79f
Signed-off-by: Daniel Cruz <dc443y@att.com>