Remove checkstyle warnings for SDC
This change intrduces a change th toe root POM that reports
checkstyle warnings. The line to prevent the build passing on checkstyle
warnings is commented out for now until all warnings have been removed.

Note: This change impacts the structure of Yaml files including
control loop definintions and heat templates because field names in
the Yaml POJOs do not comply with checkstyle rules. If impacts on the
Yaml templates are to be avoided, then this change shold be rejected
and some mechanism that overrides the checkstyle rules on those classes
must be included and approved.

Change-Id: Id8e3f159d0eca7e6d671095099e347d59df57bf6
Issue-ID: POLICY-166
Signed-off-by: liamfallon <liam.fallon@ericsson.com>