Update patch set 2
Patch Set 2: Code-Review+1

> Working ok in my tests, but have to give it a try in windriver lab
 > as soon as I see it merged. I think these changes though could
 > become very beneficial during some testing activities (runtime use
 > cases), main reason I would like them in at this stage.

I agree this would be very nice for testing. Also it looks like some changes were made to the beijing template recently that broke it and this pull request fixes it (the "getCommonHeader()" was missing for the LCM response rules) so it would be good to merge this to fix it.

Patch-set: 2
Reviewer: Gerrit User 1677 <1677@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1