Update patch set 1
Patch Set 1: Code-Review+1

Thanks Wakas, my only comment is cosmetic, we have this path (and also coordinates):  drools-applications/drools-applications-testsuite.    It seems a little bit less verbose while preserving meaning (based on parent project) to call it "testsuite" instead.

I suppose this "testsuite" module will grow with performance tests, etc .. with new projects underneath.  I think it is a little bit more readable:

drools-applications/testsuites/stability/pom.xml
drools-applications/testsuites/performance/pom.xml
...

Another comment is instead of being in the src/test/resources, may be it can be in src/main/resources, and apply an assembly to zip all these jmeter files.   The reason why is that when we release beijing, it will be nice to go the official nexus repo with the blessed version, for anyone that want to exercise it, and get the official released jmeter files.

This also works the way you have it, so I +1.   But I leave the comments there for your consideration, or perhaps future enhancements if the team thinks the comments are worth implementing.

Patch-set: 1
Label: Code-Review=+1